requirements.txt
astroid==2.4.1
backcall==0.1.0
decorator==4.4.2
dj-database-url==0.5.0
Django==2.2.8
django-debug-toolbar==2.1
django-heroku==0.3.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
drf-generators==0.4.0
gunicorn==20.0.4
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.2
pytz==2019.3
six==1.15.0
sqlparse==0.3.0
termcolor==1.1.0
toml==0.10.1
traitlets==4.3.3
wcwidth==0.2.3
whitenoise==5.1.0
wrapt==1.12.1
